Migrate srt generation code to srt_segment_output.go, add json_segment_output.go, and move srt generation files to new whisper_cpp_generator package
The segment output code needs to be unified under a single interface in the future an organized better. This is just a first step towards that.

The new json_segment_output.go code marshals segment data directly from whisper for further use (potentially as a save state for a restart at checkpoint feature)